crate with shared code
