+++ advent_of_code/aoc/2023/solutions/Cargo.toml
@@ -11,6 +11,7 @@ readme = "README.md"
 repository.workspace = true
 
 [dependencies]
+aoc-core = { path = "../../core"}
 itertools = "0.11.0"
 
 [dev-dependencies]
